refactor: add a shared layout
