+++ client/pages/shared/index.tsx
@@ -1,0 +1,9 @@
+import SharedLayout from 'components/SharedLayout';
+
+export default function SharedIndex() {
+  return (
+    <SharedLayout>
+      <p>Showing list of shared transactions</p>
+    </SharedLayout>
+  );
+}
